# src/main/java/com/chrisgya/springjpagraddle/controller/ProductController.java
package com.chrisgya.springjpagraddle.controller;

import com.chrisgya.springjpagraddle.ProductApi;
import com.chrisgya.springjpagraddle.hateoas.ProductRepresentationModelAssembler;
import com.chrisgya.springjpagraddle.model.Product;
import com.chrisgya.springjpagraddle.service.ProductService;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.RestController;

import javax.validation.Valid;
import java.util.List;

import static org.springframework.http.ResponseEntity.ok;

@RestController
public class ProductController implements ProductApi {

  private ProductService service;
  private final ProductRepresentationModelAssembler assembler;

  public ProductController(ProductService service, ProductRepresentationModelAssembler assembler) {
    this.service = service;
    this.assembler = assembler;
  }

  @Override
  public ResponseEntity<Product> getProduct(String id) {
    return service.getProduct(id).map(assembler::toModel).map(ResponseEntity::ok).orElse(ResponseEntity.notFound().build());
  }

  @Override
  public ResponseEntity<List<Product>> queryProducts(@Valid String tag, @Valid String name,
                                                     @Valid Integer page, @Valid Integer size) {
    return ok(assembler.toListModel(service.getAllProducts()));
  }
}